tox entry point to run pep8 1.4.6
The Wikimedia CI tests are hardcoded to run pep8==1.4.6. Introduce tox
to let developers easily redefine whatever command they need.

Not using flake8 since the repo does not pass pyflakes and others. But
it will be easy to switch to it later on.

Pin pep8 to 1.4.6 to match the version that has been running on
Wikimedia CI (Ubuntu Precise).
Move content of .pep8 to tox.ini section [pep8]. That is properly
recognized by pep8 1.4.6.
Ignore /.tox

Bug: T55354
Change-Id: I70e4a843dae5a7ebab6485e1abab2b124535f193